+++ tests/integration/mcp/test_mcp_callback_behavior.py
@@ -1,29 +1,56 @@
+"""Callback behaviors surfaced through MCP tools (end-to-end).
+
+Covers the full pipeline — a real Dash server via ``dash_duo`` + the MCP
+HTTP endpoint — for every callback signature variant and the surrounding
+tool-list conventions:
+
+- Positional, dict-based, and tuple-grouped ``inputs`` / ``state`` /
+  ``output`` forms.
+- ``State``, multi-output, ``PreventUpdate``-style no-output callbacks,
+  ``ctx.triggered_id``, pattern-matching (``ALL``/``MATCH``/``ALLSMALLER``).
+- Initial values: ``prevent_initial_call`` vs. initial-callback overrides.
+- Duplicate outputs (``allow_duplicate=True``) appearing as separate tools.
+- ``tools/list`` naming rules (64-char limit, uniqueness, built-ins).
+- A representative input-schema smoke test (label + DatePicker).
+- ``get_dash_component`` structured output via HTTP.
 """
-Integration tests for all Dash callback signature types.
 
-Each test verifies that:
-1. The MCP tool schema accurately reflects the callback's parameters
-2. Calling the tool with those parameters produces the expected result
-
-Assertions are derived from the callback definition, not the implementation.
-
-See: https://dash.plotly.com/flexible-callback-signatures
-"""
-
-from dash import Dash, Input, Output, State, dcc, html
+from dash import (
+    ALL,
+    ALLSMALLER,
+    MATCH,
+    Dash,
+    Input,
+    Output,
+    State,
+    ctx,
+    dcc,
+    html,
+    set_props,
+)
 
 from tests.integration.mcp.conftest import _mcp_call_tool, _mcp_tools
 
 
+# ---------------------------------------------------------------------------
+# Helpers
+# ---------------------------------------------------------------------------
+
+
 def _find_tool(tools, name):
-    return next(t for t in tools if t["name"] == name)
+    return next((t for t in tools if t["name"] == name), None)
 
 
 def _get_response(result):
     return result["result"]["structuredContent"]["response"]
 
 
-def test_positional_callback(dash_duo):
+# ---------------------------------------------------------------------------
+# Callback signatures — positional, multi-output, State, dict-based, tuples
+# ---------------------------------------------------------------------------
+
+
+def test_mcpb001_positional_callback(dash_duo):
     """Standard positional: Input("fruit", "value") → param named 'value'."""
     app = Dash(__name__)
     app.layout = html.Div(
@@ -33,8 +60,6 @@ def test_positional_callback(dash_duo):
         ]
     )
 
-    # Callback: 1 Input → 1 param named "value" (from function signature)
-    # Returns string → Output("out", "children")
     @app.callback(Output("out", "children"), Input("fruit", "value"))
     def show_fruit(value):
         return f"Selected: {value}"
@@ -48,23 +73,20 @@ def test_positional_callback(dash_duo):
     assert set(props.keys()) == {"value"}
     assert any(s.get("type") == "string" for s in props["value"]["anyOf"])
 
-    # Tool description reflects initial state
     value_desc = props["value"].get("description", "")
     assert "value: 'apple'" in value_desc
     assert "options: ['apple', 'banana']" in value_desc
 
-    # MCP tool with initial inputs matches browser
     result = _mcp_call_tool(dash_duo.server.url, "show_fruit", {"value": "apple"})
     response = _get_response(result)
     assert response["out"]["children"] == "Selected: apple"
 
-    # MCP tool with different inputs
     result = _mcp_call_tool(dash_duo.server.url, "show_fruit", {"value": "banana"})
     response = _get_response(result)
     assert response["out"]["children"] == "Selected: banana"
 
 
-def test_positional_with_state(dash_duo):
+def test_mcpb002_positional_with_state(dash_duo):
     """Positional with State: Input + State both appear as params."""
     app = Dash(__name__)
     app.layout = html.Div(
@@ -75,7 +97,6 @@ def test_positional_with_state(dash_duo):
         ]
     )
 
-    # Callback: 1 Input + 1 State → 2 params named "n_clicks" and "value"
     @app.callback(
         Output("out", "children"),
         Input("btn", "n_clicks"),
@@ -93,10 +114,8 @@ def test_positional_with_state(dash_duo):
     assert set(props.keys()) == {"n_clicks", "value"}
     assert any(s.get("type") == "number" for s in props["n_clicks"]["anyOf"])
 
-    # Tool description reflects initial state
     assert "value: 'hello'" in props["value"].get("description", "")
 
-    # MCP tool with initial inputs matches browser
     result = _mcp_call_tool(
         dash_duo.server.url, "update", {"n_clicks": None, "value": "hello"}
     )
@@ -110,7 +129,7 @@ def test_positional_with_state(dash_duo):
     assert response["out"]["children"] == "Clicked 3 with world"
 
 
-def test_multi_output_positional(dash_duo):
+def test_mcpb003_multi_output_positional(dash_duo):
     """Multi-output: returns tuple → both outputs updated in response."""
     app = Dash(__name__)
     app.layout = html.Div(
@@ -121,7 +140,6 @@ def test_multi_output_positional(dash_duo):
         ]
     )
 
-    # Callback: 1 Input → 2 Outputs via tuple return
     @app.callback(
         Output("out1", "children"),
         Output("out2", "children"),
@@ -138,18 +156,16 @@ def test_multi_output_positional(dash_duo):
     props = tool["inputSchema"]["properties"]
     assert set(props.keys()) == {"value"}
 
-    # Tool description reflects initial state
     assert "value: 'test'" in props["value"].get("description", "")
 
-    # MCP tool with initial inputs matches browser
     result = _mcp_call_tool(dash_duo.server.url, "split_case", {"value": "test"})
     response = _get_response(result)
     assert response["out1"]["children"] == "TEST"
     assert response["out2"]["children"] == "test"
 
 
-def test_dict_based_inputs_and_state(dash_duo):
-    """Dict-based: inputs=dict(trigger=...), state=dict(name=...) → dict keys are param names."""
+def test_mcpb004_dict_based_inputs_and_state(dash_duo):
+    """Dict-based: inputs=dict(trigger=...), state=dict(name=...) → dict keys are params."""
     app = Dash(__name__)
     app.layout = html.Div(
         [
@@ -159,7 +175,6 @@ def test_dict_based_inputs_and_state(dash_duo):
         ]
     )
 
-    # Callback: dict keys "trigger" and "name" become param names
     @app.callback(
         Output("out", "children"),
         inputs=dict(trigger=Input("btn", "n_clicks")),
@@ -177,7 +192,6 @@ def test_dict_based_inputs_and_state(dash_duo):
     assert set(props.keys()) == {"trigger", "name"}
     assert any(s.get("type") == "number" for s in props["trigger"]["anyOf"])
 
-    # MCP tool with initial inputs matches browser
     result = _mcp_call_tool(
         dash_duo.server.url, "greet", {"trigger": None, "name": "world"}
     )
@@ -191,7 +205,7 @@ def test_dict_based_inputs_and_state(dash_duo):
     assert response["out"]["children"] == "Hello, Dash!"
 
 
-def test_dict_based_outputs(dash_duo):
+def test_mcpb005_dict_based_outputs(dash_duo):
     """Dict-based outputs: output=dict(...) → callback returns dict, both outputs updated."""
     app = Dash(__name__)
     app.layout = html.Div(
@@ -202,7 +216,6 @@ def test_dict_based_outputs(dash_duo):
         ]
     )
 
-    # Callback: dict output keys "upper" and "lower" map to components
     @app.callback(
         output=dict(
             upper=Output("upper-out", "children"),
@@ -221,14 +234,13 @@ def test_dict_based_outputs(dash_duo):
     props = tool["inputSchema"]["properties"]
     assert set(props.keys()) == {"val"}
 
-    # MCP tool with initial inputs matches browser
     result = _mcp_call_tool(dash_duo.server.url, "transform", {"val": "hello"})
     response = _get_response(result)
     assert response["upper-out"]["children"] == "HELLO"
     assert response["lower-out"]["children"] == "hello"
 
 
-def test_mixed_input_state_in_inputs(dash_duo):
+def test_mcpb006_mixed_input_state_in_inputs(dash_duo):
     """Mixed: State inside inputs=dict alongside Input → all appear as params."""
     app = Dash(__name__)
     app.layout = html.Div(
@@ -240,7 +252,6 @@ def test_mixed_input_state_in_inputs(dash_duo):
         ]
     )
 
-    # Callback: Input and State mixed in same dict → all keys are params
     @app.callback(
         Output("out", "children"),
         inputs=dict(
@@ -261,7 +272,6 @@ def test_mixed_input_state_in_inputs(dash_duo):
     assert set(props.keys()) == {"clicks", "first", "last"}
     assert any(s.get("type") == "number" for s in props["clicks"]["anyOf"])
 
-    # MCP tool with initial inputs matches browser
     result = _mcp_call_tool(
         dash_duo.server.url,
         "full_name",
@@ -279,7 +289,7 @@ def test_mixed_input_state_in_inputs(dash_duo):
     assert response["out"]["children"] == "John Smith"
 
 
-def test_tuple_grouped_inputs(dash_duo):
+def test_mcpb007_tuple_grouped_inputs(dash_duo):
     """Tuple grouping: pair=(Input("a",...), Input("b",...)) → expands to two named params."""
     app = Dash(__name__)
     app.layout = html.Div(
@@ -290,7 +300,6 @@ def test_tuple_grouped_inputs(dash_duo):
         ]
     )
 
-    # Callback: tuple group "pair" maps to 2 deps → 2 params named pair_<id>__<prop>
     @app.callback(
         Output("out", "children"),
         inputs=dict(pair=(Input("a", "value"), Input("b", "value"))),
@@ -302,7 +311,6 @@ def test_tuple_grouped_inputs(dash_duo):
     tool = _find_tool(_mcp_tools(dash_duo.server.url), "combine")
     props = tool["inputSchema"]["properties"]
 
-    # Tuple expands: one param per dep, named with group prefix + component info
     assert set(props.keys()) == {"pair_a__value", "pair_b__value"}
     for schema in props.values():
         assert any(s.get("type") == "string" for s in schema["anyOf"])
@@ -316,7 +324,7 @@ def test_tuple_grouped_inputs(dash_duo):
     assert response["out"]["children"] == "x+y"
 
 
-def test_initial_values_from_chained_callbacks(dash_duo):
+def test_mcpb008_initial_values_from_chained_callbacks(dash_duo):
     """Querying components reflects post-initial-callback values.
 
     3-link chain: country (default "France") → update_states →
@@ -368,16 +376,13 @@ def test_initial_values_from_chained_callbacks(dash_duo):
 
     dash_duo.start_server(app)
 
-    # Tool descriptions should reflect post-initial-callback state
     tools = _mcp_tools(dash_duo.server.url)
     update_cities_tool = _find_tool(tools, "update_cities")
     state_desc = update_cities_tool["inputSchema"]["properties"]["state"].get(
         "description", ""
     )
-    # state.value was set to "Ile-de-France" by update_states initial callback
     assert "Ile-de-France" in state_desc
 
-    # state.value should be "Ile-de-France" (first state for France)
     result = _mcp_call_tool(
         dash_duo.server.url,
         "get_dash_component",
@@ -386,7 +391,6 @@ def test_initial_values_from_chained_callbacks(dash_duo):
     state_props = result["result"]["structuredContent"]["properties"]
     assert state_props["value"]["initial_value"] == "Ile-de-France"
 
-    # city.value should be "Paris" (first city for Ile-de-France)
     result = _mcp_call_tool(
         dash_duo.server.url,
         "get_dash_component",
@@ -396,7 +400,7 @@ def test_initial_values_from_chained_callbacks(dash_duo):
     assert city_props["value"]["initial_value"] == "Paris"
 
 
-def test_dict_based_reordered_state_input(dash_duo):
+def test_mcpb009_dict_based_reordered_state_input(dash_duo):
     """Dict-based callback with State before Input: call works, schema types correct.
 
     State is listed before Input in the dict. The callback should still
@@ -421,7 +425,6 @@ def test_dict_based_reordered_state_input(dash_duo):
 
     dash_duo.start_server(app)
 
-    # First: verify the callback actually works with these args
     result = _mcp_call_tool(
         dash_duo.server.url,
         "greet",
@@ -429,20 +432,23 @@ def test_dict_based_reordered_state_input(dash_duo):
     )
     assert _get_response(result)["out"]["children"] == "Hello Dash"
 
-    # Second: verify schema types match annotations
     tool = _find_tool(_mcp_tools(dash_duo.server.url), "greet")
     props = tool["inputSchema"]["properties"]
     assert props["trigger"]["type"] == "integer"
     assert props["name"]["type"] == "string"
 
-    # Third: verify each param describes the correct component
     trigger_desc = props["trigger"].get("description", "")
     assert "number of times that this element has been clicked on" in trigger_desc
     name_desc = props["name"].get("description", "")
     assert "The value of the input" in name_desc
 
 
-def test_pattern_matching_callback(dash_duo):
+# ---------------------------------------------------------------------------
+# Pattern-matching callbacks (ALL / MATCH / ALLSMALLER)
+# ---------------------------------------------------------------------------
+
+
+def test_mcpb010_pattern_matching_callback(dash_duo):
     """Pattern-matching dict IDs: tool works with correct params and results."""
     app = Dash(__name__)
     app.layout = html.Div(
@@ -469,7 +475,6 @@ def test_pattern_matching_callback(dash_duo):
     assert "first" in props
     assert "second" in props
 
-    # Verify initial output matches what the browser shows
     dash_duo.wait_for_text_to_equal("#out", "hello world")
     result = _mcp_call_tool(
         dash_duo.server.url,
@@ -479,7 +484,6 @@ def test_pattern_matching_callback(dash_duo):
     response = _get_response(result)
     assert response["out"]["children"] == "hello world"
 
-    # Verify with different values
     result = _mcp_call_tool(
         dash_duo.server.url,
         "combine",
@@ -489,10 +493,8 @@ def test_pattern_matching_callback(dash_duo):
     assert response["out"]["children"] == "foo bar"
 
 
-def test_pattern_matching_with_all_wildcard(dash_duo):
+def test_mcpb011_pattern_matching_with_all_wildcard(dash_duo):
     """ALL wildcard: one callback receives values from all matching components."""
-    from dash import ALL
-
     app = Dash(__name__)
     app.layout = html.Div(
         [
@@ -515,7 +517,6 @@ def test_pattern_matching_with_all_wildcard(dash_duo):
     tool = _find_tool(_mcp_tools(dash_duo.server.url), "summarize")
     assert tool is not None
 
-    # Schema must describe values as an array of {id, property, value} objects
     values_schema = tool["inputSchema"]["properties"]["values"]
     assert (
         values_schema["type"] == "array"
@@ -529,7 +530,6 @@ def test_pattern_matching_with_all_wildcard(dash_duo):
         "description", ""
     ), "ALL wildcard param description should explain the pattern-matching behavior"
 
-    # MCP tool call with browser-like format: concrete IDs + values
     result = _mcp_call_tool(
         dash_duo.server.url,
         "summarize",
@@ -551,7 +551,6 @@ def test_pattern_matching_with_all_wildcard(dash_duo):
     response = _get_response(result)
     assert response["summary"]["children"] == "alpha, beta"
 
-    # Different values
     result = _mcp_call_tool(
         dash_duo.server.url,
         "summarize",
@@ -574,10 +573,8 @@ def test_pattern_matching_with_all_wildcard(dash_duo):
     assert response["summary"]["children"] == "one, two"
 
 
-def test_pattern_matching_mixed_outputs(dash_duo):
+def test_mcpb012_pattern_matching_mixed_outputs(dash_duo):
     """Mixed outputs: one regular + one ALL wildcard in the same callback."""
-    from dash import ALL
-
     app = Dash(__name__)
     app.layout = html.Div(
         [
@@ -624,13 +621,11 @@ def test_pattern_matching_mixed_outputs(dash_duo):
     assert response["total"]["children"] == "Total: 2 items"
 
 
-def test_pattern_matching_with_match_wildcard(dash_duo):
+def test_mcpb013_pattern_matching_with_match_wildcard(dash_duo):
     """MATCH wildcard: callback fires per-component with matching index.
 
     Based on https://dash.plotly.com/pattern-matching-callbacks
     """
-    from dash import MATCH
-
     app = Dash(__name__)
     app.layout = html.Div(
         [
@@ -661,11 +656,9 @@ def test_pattern_matching_with_match_wildcard(dash_duo):
     tool = _find_tool(_mcp_tools(dash_duo.server.url), "show_city")
     assert tool is not None
 
-    # Schema describes MATCH input
     value_schema = tool["inputSchema"]["properties"]["value"]
     assert "Pattern-matching input (MATCH)" in value_schema.get("description", "")
 
-    # Call with concrete ID for index 0 (MATCH takes a single entry, not an array)
     result = _mcp_call_tool(
         dash_duo.server.url,
         "show_city",
@@ -678,18 +671,15 @@ def test_pattern_matching_with_match_wildcard(dash_duo):
         },
     )
     response = _get_response(result)
-    # Find the output key containing "city-out" (Dash may serialize dict IDs differently)
     out_key = next(k for k in response if "city-out" in k)
     assert response[out_key]["children"] == "Selected: MTL"
 
 
-def test_pattern_matching_with_allsmaller_wildcard(dash_duo):
+def test_mcpb014_pattern_matching_with_allsmaller_wildcard(dash_duo):
     """ALLSMALLER wildcard: receives values from components with smaller index.
 
     Based on https://dash.plotly.com/pattern-matching-callbacks
     """
-    from dash import MATCH, ALLSMALLER
-
     app = Dash(__name__)
     app.layout = html.Div(
         [
@@ -728,14 +718,12 @@ def test_pattern_matching_with_allsmaller_wildcard(dash_duo):
     tool = _find_tool(_mcp_tools(dash_duo.server.url), "show_countries")
     assert tool is not None
 
-    # Schema describes both MATCH and ALLSMALLER inputs
     props = tool["inputSchema"]["properties"]
     assert "Pattern-matching input (MATCH)" in props["current"].get("description", "")
     assert "Pattern-matching input (ALLSMALLER)" in props["previous"].get(
         "description", ""
     )
 
-    # Call for index 2: MATCH is a single dict, ALLSMALLER is a list
     result = _mcp_call_tool(
         dash_duo.server.url,
         "show_countries",
@@ -764,13 +752,13 @@ def test_pattern_matching_with_allsmaller_wildcard(dash_duo):
     assert response[out_key]["children"] == "All: Japan, Germany, France"
 
 
-def test_prevent_initial_call_uses_layout_default(dash_duo):
-    """prevent_initial_call=True: initial value stays as the layout default.
+# ---------------------------------------------------------------------------
+# Initial values: prevent_initial_call vs. initial-callback overrides
+# ---------------------------------------------------------------------------
 
-    The dropdown has value="original" in the layout. The callback has
-    prevent_initial_call=True so it doesn't run on page load. The MCP
-    tool description should show value: 'a' (layout default).
-    """
+
+def test_mcpb015_prevent_initial_call_uses_layout_default(dash_duo):
+    """prevent_initial_call=True: initial value stays as the layout default."""
     app = Dash(__name__)
     app.layout = html.Div(
         [
@@ -788,24 +776,16 @@ def test_prevent_initial_call_uses_layout_default(dash_duo):
         return f"Changed to: {val}"
 
     dash_duo.start_server(app)
-    # Browser shows layout default — callback hasn't fired
     dash_duo.wait_for_text_to_equal("#out", "not yet")
 
     tool = _find_tool(_mcp_tools(dash_duo.server.url), "update")
     val_desc = tool["inputSchema"]["properties"]["val"].get("description", "")
 
-    # Tool description reflects layout default, not callback output
     assert "value: 'a'" in val_desc
 
 
-def test_initial_callback_overrides_layout_value(dash_duo):
-    """Initial callback overrides layout value in tool description.
-
-    The city dropdown has value="default-city" in the layout.
-    update_city runs on page load (no prevent_initial_call) and
-    sets city.value to "Paris". The MCP tool should show "Paris"
-    as the default, not "default-city".
-    """
+def test_mcpb016_initial_callback_overrides_layout_value(dash_duo):
+    """Initial callback overrides layout value in tool description."""
     app = Dash(__name__)
     app.layout = html.Div(
         [
@@ -830,24 +810,20 @@ def test_initial_callback_overrides_layout_value(dash_duo):
         return f"City: {city}"
 
     dash_duo.start_server(app)
-    # Browser shows "Paris" — the initial callback overrode "default-city"
     dash_duo.wait_for_text_to_equal("#out", "City: Paris")
 
     tool = _find_tool(_mcp_tools(dash_duo.server.url), "show_city")
     city_desc = tool["inputSchema"]["properties"]["city"].get("description", "")
 
-    # Tool description should show the post-initial-callback value
     assert "value: 'Paris'" in city_desc
     assert "default-city" not in city_desc
 
 
-def test_callback_context_triggered_id(dash_duo):
+def test_mcpb017_callback_context_triggered_id(dash_duo):
     """Callbacks using dash.ctx.triggered_id work via MCP.
 
     Based on https://dash.plotly.com/determining-which-callback-input-changed
     """
-    from dash import ctx
-
     app = Dash(__name__)
     app.layout = html.Div(
         [
@@ -871,17 +847,14 @@ def test_callback_context_triggered_id(dash_duo):
 
     dash_duo.start_server(app)
 
-    # Browser initial state: no button clicked
     dash_duo.wait_for_text_to_equal("#output", "No button clicked yet")
 
-    # Tool should have all three button params
     tool = _find_tool(_mcp_tools(dash_duo.server.url), "display")
     props = tool["inputSchema"]["properties"]
     assert "btn1" in props
     assert "btn2" in props
     assert "btn3" in props
 
-    # Click btn-2 via MCP — ctx.triggered_id should be "btn-2"
     result = _mcp_call_tool(
         dash_duo.server.url,
         "display",
@@ -890,7 +863,6 @@ def test_callback_context_triggered_id(dash_duo):
     response = _get_response(result)
     assert response["output"]["children"] == "Last clicked: btn-2"
 
-    # Click btn-3 via MCP
     result = _mcp_call_tool(
         dash_duo.server.url,
         "display",
@@ -900,14 +872,12 @@ def test_callback_context_triggered_id(dash_duo):
     assert response["output"]["children"] == "Last clicked: btn-3"
 
 
-def test_no_output_callback_does_not_crash_tools_list(dash_duo):
+def test_mcpb018_no_output_callback_does_not_crash_tools_list(dash_duo):
     """A callback with no Output should not crash tools/list.
 
     No-output callbacks use set_props for side effects. They produce
     a hash-only output_id with no dot separator.
     """
-    from dash import set_props
-
     app = Dash(__name__)
     app.layout = html.Div(
         [
@@ -930,13 +900,9 @@ def test_no_output_callback_does_not_crash_tools_list(dash_duo):
     tools = _mcp_tools(dash_duo.server.url)
     tool_names = [t["name"] for t in tools]
 
-    # show_selection should appear as a tool
     assert "show_selection" in tool_names
-
-    # log_click has no declared output but uses set_props — still a valid tool
     assert "log_click" in tool_names
 
-    # Call log_click — sideUpdate should show the set_props effect
     result = _mcp_call_tool(
         dash_duo.server.url,
         "log_click",
@@ -946,9 +912,6 @@ def test_no_output_callback_does_not_crash_tools_list(dash_duo):
     assert "sideUpdate" in structured
     assert structured["sideUpdate"]["display"]["children"] == "Logged 3 clicks"
 
-    # get_dash_component shows show_selection as modifier (declared output).
-    # log_click uses set_props which bypasses the declarative graph —
-    # its effect is only visible via sideUpdate in tool call results.
     result = _mcp_call_tool(
         dash_duo.server.url,
         "get_dash_component",
@@ -956,3 +919,338 @@ def test_no_output_callback_does_not_crash_tools_list(dash_duo):
     )
     prop_info = result["result"]["structuredContent"]["properties"]["children"]
     assert "show_selection" in prop_info["modified_by_tool"]
+
+
+# ---------------------------------------------------------------------------
+# Duplicate outputs (allow_duplicate=True)
+# ---------------------------------------------------------------------------
+
+
+def test_mcpb019_duplicate_outputs_both_tools_listed(dash_duo):
+    """Both callbacks outputting to the same component appear as tools."""
+    app = Dash(__name__)
+    app.layout = html.Div(
+        [
+            dcc.Input(id="first-name", value="Jane"),
+            dcc.Input(id="last-name", value="Doe"),
+            html.Div(id="greeting"),
+        ]
+    )
+
+    @app.callback(
+        Output("greeting", "children"),
+        Input("first-name", "value"),
+    )
+    def greet_by_first(first):
+        return f"Hello, {first}!"
+
+    @app.callback(
+        Output("greeting", "children", allow_duplicate=True),
+        Input("last-name", "value"),
+        prevent_initial_call=True,
+    )
+    def greet_by_last(last):
+        return f"Hi, {last}!"
+
+    dash_duo.start_server(app)
+    tools = _mcp_tools(dash_duo.server.url)
+
+    first_tool = _find_tool(tools, "greet_by_first")
+    last_tool = _find_tool(tools, "greet_by_last")
+
+    assert first_tool is not None, "greet_by_first should be listed"
+    assert last_tool is not None, "greet_by_last should be listed"
+
+
+def test_mcpb020_duplicate_outputs_both_callable(dash_duo):
+    """Both callbacks can be called and produce correct results."""
+    app = Dash(__name__)
+    app.layout = html.Div(
+        [
+            dcc.Input(id="first-name", value="Jane"),
+            dcc.Input(id="last-name", value="Doe"),
+            html.Div(id="greeting"),
+        ]
+    )
+
+    @app.callback(
+        Output("greeting", "children"),
+        Input("first-name", "value"),
+    )
+    def greet_by_first(first):
+        return f"Hello, {first}!"
+
+    @app.callback(
+        Output("greeting", "children", allow_duplicate=True),
+        Input("last-name", "value"),
+        prevent_initial_call=True,
+    )
+    def greet_by_last(last):
+        return f"Hi, {last}!"
+
+    dash_duo.start_server(app)
+
+    result1 = _mcp_call_tool(dash_duo.server.url, "greet_by_first", {"first": "Alice"})
+    assert _get_response(result1)["greeting"]["children"] == "Hello, Alice!"
+
+    result2 = _mcp_call_tool(dash_duo.server.url, "greet_by_last", {"last": "Smith"})
+    assert _get_response(result2)["greeting"]["children"] == "Hi, Smith!"
+
+
+def test_mcpb021_duplicate_outputs_find_by_output_returns_primary(dash_duo):
+    """find_by_output returns the primary (non-duplicate) callback."""
+    app = Dash(__name__)
+    app.layout = html.Div(
+        [
+            dcc.Input(id="first-name", value="Jane"),
+            dcc.Input(id="last-name", value="Doe"),
+            html.Div(id="greeting"),
+        ]
+    )
+
+    @app.callback(
+        Output("greeting", "children"),
+        Input("first-name", "value"),
+    )
+    def greet_by_first(first):
+        return f"Hello, {first}!"
+
+    @app.callback(
+        Output("greeting", "children", allow_duplicate=True),
+        Input("last-name", "value"),
+        prevent_initial_call=True,
+    )
+    def greet_by_last(last):
+        return f"Hi, {last}!"
+
+    dash_duo.start_server(app)
+
+    result = _mcp_call_tool(
+        dash_duo.server.url,
+        "get_dash_component",
+        {"component_id": "greeting", "property": "children"},
+    )
+    structured = result["result"]["structuredContent"]
+    assert structured["properties"]["children"]["initial_value"] == "Hello, Jane!"
+
+
+# ---------------------------------------------------------------------------
+# tools/list — naming rules (64-char limit, uniqueness, built-ins)
+# ---------------------------------------------------------------------------
+
+
+def test_mcpb022_tool_names_within_64_chars(dash_duo):
+    app = Dash(__name__)
+    app.layout = html.Div(
+        [
+            dcc.Dropdown(id="dd", options=["a"], value="a"),
+            html.Div(id="out"),
+        ]
+    )
+
+    @app.callback(Output("out", "children"), Input("dd", "value"))
+    def update(val):
+        return val
+
+    dash_duo.start_server(app)
+    for tool in _mcp_tools(dash_duo.server.url):
+        assert len(tool["name"]) <= 64, f"Tool name exceeds 64 chars: {tool['name']}"
+        for param_name in tool.get("inputSchema", {}).get("properties", {}):
+            assert len(param_name) <= 64, f"Param name exceeds 64 chars: {param_name}"
+
+
+def test_mcpb023_long_callback_ids_within_64_chars(dash_duo):
+    app = Dash(__name__)
+    long_id = "a" * 120
+    app.layout = html.Div(
+        [
+            dcc.Input(id=long_id, value="test"),
+            html.Div(id=f"{long_id}-output"),
+        ]
+    )
+
+    @app.callback(Output(f"{long_id}-output", "children"), Input(long_id, "value"))
+    def process(val):
+        return val
+
+    dash_duo.start_server(app)
+    for tool in _mcp_tools(dash_duo.server.url):
+        assert len(tool["name"]) <= 64, f"Tool name exceeds 64 chars: {tool['name']}"
+
+
+def test_mcpb024_pattern_matching_ids_within_64_chars(dash_duo):
+    app = Dash(__name__)
+    app.layout = html.Div(
+        [
+            html.Div(
+                [
+                    dcc.Input(
+                        id={"type": "filter-input", "index": i, "category": "primary"},
+                        value=f"val-{i}",
+                    )
+                    for i in range(3)
+                ]
+            ),
+            html.Div(id="pm-output"),
+        ]
+    )
+
+    @app.callback(
+        Output("pm-output", "children"),
+        Input({"type": "filter-input", "index": 0, "category": "primary"}, "value"),
+    )
+    def filter_update(v0):
+        return str(v0)
+
+    dash_duo.start_server(app)
+    for tool in _mcp_tools(dash_duo.server.url):
+        assert len(tool["name"]) <= 64, f"Tool name exceeds 64 chars: {tool['name']}"
+
+
+def test_mcpb025_duplicate_func_names_produce_unique_tools(dash_duo):
+    app = Dash(__name__)
+    app.layout = html.Div(
+        [
+            dcc.Dropdown(id="dd1", options=["a"], value="a"),
+            html.Div(id="dd1-output"),
+            dcc.Dropdown(id="dd2", options=["b"], value="b"),
+            html.Div(id="dd2-output"),
+            dcc.Dropdown(id="dd3", options=["c"], value="c"),
+            html.Div(id="dd3-output"),
+        ]
+    )
+
+    @app.callback(Output("dd1-output", "children"), Input("dd1", "value"))
+    def cb(value):
+        return f"first: {value}"
+
+    @app.callback(Output("dd2-output", "children"), Input("dd2", "value"))
+    def cb(value):  # noqa: F811
+        return f"second: {value}"
+
+    @app.callback(Output("dd3-output", "children"), Input("dd3", "value"))
+    def cb(value):  # noqa: F811
+        return f"third: {value}"
+
+    dash_duo.start_server(app)
+    tools = _mcp_tools(dash_duo.server.url)
+    cb_tools = [t for t in tools if t["name"] not in ("get_dash_component",)]
+    tool_names = [t["name"] for t in cb_tools]
+
+    assert (
+        len(tool_names) == 3
+    ), f"Expected 3 callback tools, got {len(tool_names)}: {tool_names}"
+    assert len(set(tool_names)) == 3, f"Tool names not unique: {tool_names}"
+
+
+def test_mcpb026_builtin_tools_always_present(dash_duo):
+    app = Dash(__name__)
+    app.layout = html.Div(id="root")
+
+    dash_duo.start_server(app)
+    tool_names = [t["name"] for t in _mcp_tools(dash_duo.server.url)]
+    assert "get_dash_component" in tool_names
+
+
+# ---------------------------------------------------------------------------
+# Input schema smoke test + get_dash_component HTTP structured output
+# ---------------------------------------------------------------------------
+
+
+def test_mcpb027_mcp_tool_with_label_and_date_picker_schema(dash_duo):
+    """Full assertion on a tool with an html.Label and DatePickerSingle constraints."""
+    label_text = "Departure Date"
+    component_id = "dp"
+    min_date = "2020-01-01"
+    max_date = "2025-12-31"
+    default_date = "2024-06-15"
+    func_name = "select_date"
+    param_name = "date"  # function parameter name
+
+    app = Dash(__name__)
+    app.layout = html.Div(
+        [
+            html.Label(label_text, htmlFor=component_id),
+            dcc.DatePickerSingle(
+                id=component_id,
+                min_date_allowed=min_date,
+                max_date_allowed=max_date,
+                date=default_date,
+            ),
+            html.Div(id="out"),
+        ]
+    )
+
+    @app.callback(Output("out", "children"), Input(component_id, "date"))
+    def select_date(date):
+        return f"Selected: {date}"
+
+    dash_duo.start_server(app)
+    tools = _mcp_tools(dash_duo.server.url)
+
+    tool = next(t for t in tools if t["name"] not in ("get_dash_component",))
+
+    assert func_name in tool["name"]
+
+    schema = tool["inputSchema"]
+    assert schema["type"] == "object"
+    assert param_name in schema["required"]
+    assert param_name in schema["properties"]
+
+    prop = schema["properties"][param_name]
+    assert prop["type"] == "string"
+    assert prop["format"] == "date"
+
+    desc = prop["description"]
+    for expected in (label_text, min_date, max_date, default_date):
+        assert expected in desc, f"Expected {expected!r} in description: {desc!r}"
+
+
+EXPECTED_DROPDOWN_OPTIONS = {
+    "component_id": "my-dropdown",
+    "component_type": "Dropdown",
+    "label": None,
+    "properties": {
+        "options": {
+            "initial_value": [
+                {"label": "New York", "value": "NYC"},
+                {"label": "Montreal", "value": "MTL"},
+            ],
+            "modified_by_tool": [],
+            "input_to_tool": [],
+        },
+    },
+}
+
+
+def test_mcpb028_query_component_returns_structured_output(dash_duo):
+    app = Dash(__name__)
+    app.layout = html.Div(
+        [
+            dcc.Dropdown(
+                id="my-dropdown",
+                options=[
+                    {"label": "New York", "value": "NYC"},
+                    {"label": "Montreal", "value": "MTL"},
+                ],
+                value="NYC",
+            ),
+        ]
+    )
+
+    dash_duo.start_server(app)
+
+    result = _mcp_call_tool(
+        dash_duo.server.url,
+        "get_dash_component",
+        {"component_id": "my-dropdown", "property": "options"},
+    )
+
+    assert "result" in result, f"Expected result in response: {result}"
+    structured = result["result"]["structuredContent"]
+    assert structured["component_id"] == EXPECTED_DROPDOWN_OPTIONS["component_id"]
+    assert structured["component_type"] == EXPECTED_DROPDOWN_OPTIONS["component_type"]
+    assert (
+        structured["properties"]["options"]
+        == EXPECTED_DROPDOWN_OPTIONS["properties"]["options"]
+    )
